@@ -1,25 +1,12 @@
 import React from "react";
-import logo from "./logo.svg";
 import "./App.css";
+
+import AutocompleteEditor from "./components/AutocompleteEditor";
 
 function App() {
   return (
     <div className="App">
-      <header className="App-header">
-        <img src={logo} className="App-logo" alt="logo" />
-        <p>
-          Edit <code>src/App.js</code> and save to reload. Please write the
-          autocomplete here.
-        </p>
-        <a
-          className="App-link"
-          href="https://reactjs.org"
-          target="_blank"
-          rel="noopener noreferrer"
-        >
-          Learn React
-        </a>
-      </header>
+      <AutocompleteEditor />
     </div>
   );
 }
